GUI: win32 simulation fix gcc compile

+++ realgui/input/tp_algo.c
@@ -262,7 +262,7 @@ static bool tp_judge_slow_x_left_slide(struct gui_touch_port_data *raw_data)
     (void)raw_data;
     if (tp.type == TOUCH_HOLD_X)
     {
-        if (abs(tp.deltaX) > gui_get_screen_width() / 2)
+        if (abs(tp.deltaX) > (int)(gui_get_screen_width() / 2))
         {
             if (tp.deltaX < 0)
             {
@@ -279,7 +279,7 @@ static bool tp_judge_slow_x_right_slide(struct gui_touch_port_data *raw_data)
     (void)raw_data;
     if (tp.type == TOUCH_HOLD_X)
     {
-        if (abs(tp.deltaX) > gui_get_screen_width() / 2)
+        if (abs(tp.deltaX) > (int)(gui_get_screen_width() / 2))
         {
             if (tp.deltaX > 0)
             {
@@ -296,7 +296,7 @@ static bool tp_judge_slow_x_orign_slide(struct gui_touch_port_data *raw_data)
     (void)raw_data;
     if (tp.type == TOUCH_HOLD_X)
     {
-        if (abs(tp.deltaX) <= gui_get_screen_width() / 2)
+        if (abs(tp.deltaX) <= (int)(gui_get_screen_width() / 2))
         {
             tp.type = TOUCH_ORIGIN_FROM_X;
             TP_LOG("type = TOUCH_ORIGIN_FROM_X\n");
@@ -310,7 +310,7 @@ static bool tp_judge_slow_y_down_slide(struct gui_touch_port_data *raw_data)
     (void)raw_data;
     if (tp.type == TOUCH_HOLD_Y)
     {
-        if (abs(tp.deltaY) > gui_get_screen_height() / 2)
+        if (abs(tp.deltaY) > (int)(gui_get_screen_height() / 2))
         {
             if (tp.deltaY > 0)
             {
@@ -327,7 +327,7 @@ static bool tp_judge_slow_y_up_slide(struct gui_touch_port_data *raw_data)
     (void)raw_data;
     if (tp.type == TOUCH_HOLD_Y)
     {
-        if (abs(tp.deltaY) > gui_get_screen_height() / 2)
+        if (abs(tp.deltaY) > (int)(gui_get_screen_height() / 2))
         {
             if (tp.deltaY < 0)
             {
@@ -344,7 +344,7 @@ static bool tp_judge_slow_y_orign_slide(struct gui_touch_port_data *raw_data)
     (void)raw_data;
     if (tp.type == TOUCH_HOLD_Y)
     {
-        if (abs(tp.deltaY) <= gui_get_screen_height() / 2)
+        if (abs(tp.deltaY) <= (int)(gui_get_screen_height() / 2))
         {
             tp.type = TOUCH_ORIGIN_FROM_Y;
             TP_LOG("type = TOUCH_ORIGIN_FROM_Y\n");
